refactor(hir): split class member-registration helpers out of class_decl.rs
class_decl.rs was at 1989 lines; #9315's additions take it over the 2000-line
gate. The non-computed member registration and Symbol.iterator wrapper helpers
move to a sibling module unchanged.
